Add files via upload
The advancement of computational biology is used in the healthcare industry. It allowed collecting the stored patient data for the medical disease prediction.  Healthcare is the most prime example of how machine learning is use in medical field.There are different intelligent prediction algorithms are available for the diagnosis of the disease at early stages.For diagnosis entire medical history and physical tests are used. As these tests produces large amount of data and ML can be used for finding important features from large amount of data. Due to this specialty of ML can be used in combination with medical science for the accurate diagnosis of hypo thyroidism disease. A number of ML techniques have been evolved and in order to achieve best accuracy of a model ensembles are widely used.